backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-tavily
langchain-text-splitters
langgraph
chromadb
motor
pydantic
python-dotenv
sse-starlette
pymupdf
python-multipart
scikit-learn
numpy
plotly
pinecone
